+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 Change Log
 ==========
 
+0.3.0
+-----
+*2016-06-14*
+
+- ([#43](https://github.com/trello/victor/pull/43)) Experimental SVG -> Android drawable support
+
+0.2.0
+-----
+*2016-03-24*
+
+- ([#40](https://github.com/trello/victor/pull/40)) Fixed some codec issues with Batik 1.8
+- ([#38](https://github.com/trello/victor/pull/38)) Support Gradle 2.12
+
+0.1.5
+-----
+*2015-08-07*
+
+- (#28) Upgraded underlying rasterization lib to batik 1.8
+
 0.1.4
 -----
 *2015-05-07*